Rename json-fs to fs-wrapper
