+++ lib/autofix.js
@@ -4,7 +4,7 @@ const chalk = require('chalk');
 const prompts = require('prompts');
 const spawn = require('cross-spawn');
 const getPathKey = require('path-key');
-const FS = require('./json-fs');
+const FS = require('./fs-wrapper');
 const AppState = require('./state');
 const ErrorMessage = require('./error-message');
 const promisifyPort = require('./promisify-port');
