+++ lib/build.js
@@ -8,8 +8,8 @@ const wrap = require('wrap-ansi');
 const {hashElement} = require('folder-hash');
 const elmCompiler = require('../vendor/node-elm-compiler');
 const {getElmBinary} = require('./elm-binary');
-const FS = require('./json-fs');
 const Debug = require('./debug');
+const FS = require('./fs-wrapper');
 const Spinner = require('./spinner');
 const Anonymize = require('./anonymize');
 const Benchmark = require('./benchmark');
