+++ lib/cache.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const util = require('util');
 const fs = require('fs-extra');
-const FS = require('./json-fs');
+const FS = require('./fs-wrapper');
 const AppState = require('./state');
 
 const fsEnsureDir = util.promisify(fs.ensureDir);
