+++ lib/elm-files.js
@@ -4,8 +4,8 @@ const glob = require('glob');
 const fs = require('fs-extra');
 const chalk = require('chalk');
 const Hash = require('./hash');
-const FS = require('./json-fs');
 const Debug = require('./debug');
+const FS = require('./fs-wrapper');
 const Benchmark = require('./benchmark');
 const Anonymize = require('./anonymize');
 const elmParser = require('./parse-elm');
