+++ lib/fs-wrapper.js
