+++ lib/optimize-js.js
@@ -1,4 +1,4 @@
-const FS = require('./json-fs');
+const FS = require('./fs-wrapper');
 const Benchmark = require('./benchmark');
 
 module.exports = {optimize};
