+++ lib/project-json-files.js
@@ -3,7 +3,7 @@ const util = require('util');
 const os = require('os');
 const got = require('got');
 const fs = require('fs-extra');
-const FS = require('./json-fs');
+const FS = require('./fs-wrapper');
 
 const fsMkdirp = util.promisify(fs.mkdirp);
 const fsWriteJson = util.promisify(fs.writeJson);
