+++ lib/suppressed-errors.js
@@ -4,7 +4,7 @@ const childProcess = require('child_process');
 const fs = require('fs-extra');
 const glob = require('glob');
 const chalk = require('chalk');
-const FS = require('./json-fs');
+const FS = require('./fs-wrapper');
 const OsHelpers = require('./os-helpers');
 const ErrorMessage = require('./error-message');
