+++ lib/template-dependencies.js
@@ -6,8 +6,8 @@ const util = require('util');
 const fs = require('fs-extra');
 const spawnAsync = require('cross-spawn');
 const Hash = require('./hash');
-const FS = require('./json-fs');
 const Cache = require('./cache');
+const FS = require('./fs-wrapper');
 const ErrorMessage = require('./error-message');
 const getExecutable = require('elm-tooling/getExecutable');
