+++ lib/watch.js
@@ -1,12 +1,12 @@
 const path = require('path');
 const chokidar = require('chokidar');
-const FS = require('./json-fs');
 const Debug = require('./debug');
-const Anonymize = require('./anonymize');
-const AppState = require('./state');
+const FS = require('./fs-wrapper');
 const Runner = require('./runner');
-const SuppressedErrors = require('./suppressed-errors');
+const AppState = require('./state');
+const Anonymize = require('./anonymize');
 const OsHelpers = require('./os-helpers');
+const SuppressedErrors = require('./suppressed-errors');
 
 let isFlushingStdio = false;
